requirements.txt
Flask==0.10.1
Pillow==2.7.0
pytesseract==0.1.5
requests==2.6.0
Flask-Script==2.0.5
